rename: util.js -> generate-url.js
