+++ src/main/resources/static/js/meeting/list/generate-url.js
@@ -4,17 +4,14 @@ import state from "./state.js";
  * 페이징 요청 url을 리턴
  * @param {boolean} isScrollEvent 스크롤 이벤트 여부
  */
-export function generateUrl(isScrollEvent) {
-  const { isToggleOn, searchKeyword, regions, category } = state;
+export default function (isScrollEvent) {
+  const { isToggleOn, regions, category } = state;
   const meetings = document.querySelector("#meetings");
   const parameter = {};
   let lastMeetingId;
 
   // 토글 checked === true -> 모집 중인 모임만 보기
   if (isToggleOn) parameter.isClosed = false;
-
-  // 사용자가 검색한 검색어가 있을경우
-  if (searchKeyword) parameter.keyword = searchKeyword;
 
   // 지역 옵션이 선택되었다면
   if (state.regions.length !== 0) parameter.regions = regions;
@@ -29,7 +26,7 @@ export function generateUrl(isScrollEvent) {
   }
   
   const searchParams = new URLSearchParams(parameter).toString();
-
+  console.log(`/meeting/api/list?${searchParams}`);
   return `/meeting/api/list?${searchParams}`;
 }
