+++ src/main/resources/static/js/meeting/list/on-click-category.js
@@ -1,7 +1,7 @@
 import state from "./state.js";
 import requestMeetings from "./request-meetings.js";
 import { insertMeetings, removeMeetings } from "./view.js";
-import { generateUrl } from "./util.js";
+import generateUrl from "./generate-url.js";
 
 export default function(e) {
   const categoryId = e.target.dataset.id;
